# Add performance tests