--- a/mix.exs
+++ b/mix.exs
@@ -63,6 +63,7 @@ defmodule ExKeccak.MixProject do
   defp deps do
     [
       {:ex_doc, "~> 0.21", only: :dev, runtime: false},
+      {:benchee, "~> 1.0.1", only: :test},
       {:rustler, "~> 0.21.1"}
     ]
   end