@tzumby 2020-09-14
Thanks! I suspect it has something to do with how we allocate the mutable binary for returning the data in Rust.